SOLR-16203: Refactor inform to IndexSchema class
